ai-agent-demo_final/requirements.txt

# Core
python-dotenv>=1.0.1
pydantic>=2.7.0
requests>=2.31.0
tqdm>=4.66.0

# LLM / Agents
smolagents>=0.2.0
langchain>=0.2.0
langchain-community>=0.2.0
litellm>=1.43.6

# Vector DB / RAG
chromadb>=0.5.0
faiss-cpu>=1.7.4
unstructured>=0.14.0
pypdf>=4.2.0
python-docx>=1.1.2
sentence-transformers>=2.6.0

# Web
streamlit>=1.34.0
fastapi>=0.111.0
uvicorn>=0.30.0

# Cache / Storage
redis>=5.0.4

# Testing
pytest>=8.2.0
